requirements.txt
aiohttp~=3.7.4.post0
motor~=3.0.0
termcolor~=1.1.0
discord.py @ git+https://github.com/Rapptz/discord.py
